Report random seed with verbosity 1 or higher
If the test run is executing in random order, this adds the random seed number to the message printed at the beginning of the session so that testers can reproduce the same order.
It also changes this information (which is just a single line) to be reported at the "stats" level, rather than at the most-verbose "listAll" level.